# Replace ElasticArray with Vec