--- a/Cargo.toml
+++ b/Cargo.toml
@@ -7,6 +7,5 @@ version = "0.3.0"
 authors = ["CodeChain Team <hi@codechain.io>", "Parity Technologies <admin@parity.io>"]
 
 [dependencies]
-elastic-array = "0.10"
 primitives = { git = "https://github.com/CodeChain-io/rust-codechain-primitives.git", version = "0.4" }
 rustc-hex = "1.0"