ElasticArray doesn't help the performance in most cases.

before:
```
test bench_decode_1000_values        ... bench:      53,100 ns/iter (+/- 3,370)
test bench_decode_nested_empty_lists ... bench:         273 ns/iter (+/- 19)
test bench_decode_u256_value         ... bench:          26 ns/iter (+/- 2)
test bench_decode_u64_value          ... bench:          30 ns/iter (+/- 1)
test bench_stream_1000_empty_lists   ... bench:       9,574 ns/iter (+/- 624)
test bench_stream_nested_empty_lists ... bench:         213 ns/iter (+/- 4)
test bench_stream_u256_value         ... bench:         397 ns/iter (+/- 21)
test bench_stream_u64_value          ... bench:         116 ns/iter (+/- 9)
```

after:
```
test bench_decode_1000_values        ... bench:      52,146 ns/iter (+/- 1,851)
test bench_decode_nested_empty_lists ... bench:         278 ns/iter (+/- 9)
test bench_decode_u256_value         ... bench:          28 ns/iter (+/- 1)
test bench_decode_u64_value          ... bench:          25 ns/iter (+/- 1)
test bench_stream_1000_empty_lists   ... bench:       6,256 ns/iter (+/- 607)
test bench_stream_nested_empty_lists ... bench:          92 ns/iter (+/- 6)
test bench_stream_u256_value         ... bench:         316 ns/iter (+/- 19)
test bench_stream_u64_value          ... bench:          36 ns/iter (+/- 3)
```